+++ GeneratorExe/FileIO/AssemblyInfoFileAccessor.cs
@@ -96,7 +96,7 @@ namespace FileIO {
     public void WritePackageDependencies(
       DependencyInfo[] packageDependencies,
       bool addNew, bool updateExisiting, bool deleteOthers,
-      string onlyForTargetFramework
+      bool allowDowngrade, string onlyForTargetFramework
     ) { 
     
     }