+++ GeneratorExe/FileIO/PackagesConfigFileAccessor.cs
@@ -79,7 +79,7 @@ namespace FileIO {
     public void WritePackageDependencies(
       DependencyInfo[] packageDependencies,
       bool addNew, bool updateExisiting, bool deleteOthers,
-      string onlyForTargetFramework
+      bool allowDowngrade, string onlyForTargetFramework
     ) {
 
       DependencyUpdateHelper updateHelper = new DependencyUpdateHelper(
@@ -87,14 +87,14 @@ namespace FileIO {
       );
 
       updateHelper.WritePackageDependencies(
-        packageDependencies, addNew, updateExisiting, deleteOthers, onlyForTargetFramework
+        packageDependencies, addNew, updateExisiting, deleteOthers, allowDowngrade, onlyForTargetFramework
       );
 
       if(_ProjectFileAccessor != null) {
 
         //wir als führende einheit schieben zusätzlich zurück ins parent-projekt...
         _ProjectFileAccessor._LocalUpdateHelper.WritePackageDependencies(
-          packageDependencies, addNew, updateExisiting, deleteOthers, onlyForTargetFramework
+          packageDependencies, addNew, updateExisiting, deleteOthers, allowDowngrade, onlyForTargetFramework
         );
 
       }